feat(create-study-item,study-item): add create-study-item feature and progress stepper
- Add features/create-study-item: useCreateStudyItem hook, useStudyItemForm, default form values, StudyItemForm and StudyItemDrawer UI
- Add progress stepper UI under entities/study-item/ui/progress-steps (ProgressStepper, StudyItemProgressCard)
- Implement createStepTimeline in entities/study-item model utils and re-export from model/ui/index
- Remove old study-item-management/server utils and update widgets to use new create-study-item exports and mutation API (mutate/isPending)
