+++ src/entities/study-item/index.tsx
@@ -1,1 +1,3 @@
 export { useStudyItem } from "./model";
+export { createStepTimeline } from "./model";
+export { ProgressStepper, StudyItemProgressCard } from "./ui";
