+++ src/entities/study-item/model/index.tsx
@@ -1,2 +1,2 @@
-export { canRowExpand, formatCreatedDate } from "./utils";
+export { canRowExpand, formatCreatedDate, createStepTimeline } from "./utils";
 export { useStudyItem } from "./use-study-item";
