+++ src/entities/study-item/model/utils.tsx
@@ -1,6 +1,8 @@
 import type { StudyItem } from "@prisma/client";
 import type { createFormatter } from "next-intl";
-import type { TimeTranslations } from "@/shared/types";
+
+import { generateRepetitionSchedule } from "@/shared/lib/utils";
+import type { IntlFormatter, TimeTranslations } from "@/shared/types";
 
 export const canRowExpand = (item: StudyItem) => Boolean(item.description);
 
@@ -36,4 +38,38 @@ export function formatCreatedDate(
   format: ReturnType<typeof createFormatter>,
 ) {
   return format.dateTime(date, { dateStyle: "short" });
+}
+
+export function createStepTimeline(format: IntlFormatter, t: TimeTranslations) {
+  const formatStepDate = (date: Date): string => {
+    return format.dateTime(date, {
+      hour: "2-digit",
+      minute: "2-digit",
+      day: "2-digit",
+      month: "short",
+      year: "numeric",
+    });
+  };
+
+  const formatStepDateTooltip = (date: Date): string => {
+    return format.dateTime(date, {
+      weekday: "long",
+      day: "2-digit",
+      month: "long",
+      year: "numeric",
+      hour: "2-digit",
+      minute: "2-digit",
+      second: "2-digit",
+    });
+  };
+
+  return generateRepetitionSchedule().map((rep, idx, arr) => {
+    const prev = idx > 0 ? arr[idx - 1]!.scheduledAt : null;
+    return {
+      step: rep.repetitionNumber,
+      date: formatStepDate(rep.scheduledAt),
+      diff: prev ? formatDiff(t, prev, rep.scheduledAt) : "",
+      tooltip: formatStepDateTooltip(rep.scheduledAt),
+    };
+  });
 }
