+++ src/entities/study-item/ui/index.tsx
@@ -6,3 +6,5 @@ export {
   RepetitionsCell,
   HeaderCell,
 } from "./study-item-cell";
+
+export { ProgressStepper, StudyItemProgressCard } from "./progress-steps";
