+++ src/entities/study-item/ui/progress-steps/index.tsx
@@ -1,0 +1,2 @@
+export { ProgressStepper } from "./study-item-progress";
+export { StudyItemProgressCard } from "./study-item-progress-card";
