+++ src/entities/study-item/ui/progress-steps/study-item-progress-card.tsx
