+++ src/entities/study-item/ui/progress-steps/study-item-progress.tsx
@@ -15,8 +15,7 @@ import {
   TooltipContent,
   TooltipTrigger,
 } from "@/components/ui/tooltip";
-
-import { generateStepsServer } from "../model/utils";
+import { createStepTimeline } from "@/entities/study-item";
 
 type BadgeContentProps = {
   diff: string | null;
@@ -31,7 +30,7 @@ const BadgeContent: FC<BadgeContentProps> = ({ diff }) => {
 export const ProgressStepper = () => {
   const format = useFormatter();
   const t = useTranslations("Time");
-  const steps = generateStepsServer(format, t);
+  const steps = createStepTimeline(format, t);
   return (
     <div className="!w-full space-y-8 text-start">
       <Stepper defaultValue={1} orientation="vertical" className="!w-full">
