+++ src/features/create-study-item/model/utils.ts
@@ -1,0 +1,14 @@
+import { type CreateStudyItemInput } from "@/shared/api/schemas";
+
+export const defaultStudyItemFormValues: Partial<CreateStudyItemInput> = {
+  title: "",
+  description: "",
+  tagIds: [],
+};
+
+export type StepInfo = {
+  step: number;
+  date: string;
+  diff: string;
+  tooltip: string;
+};
